Modify(jiwon/signup): 회원가입 시 전공 컴포넌트 분해 #13
