+++ app/src/screens/auth/Signup.js
@@ -14,113 +14,7 @@ import {
   checkPassword,
   checkName,
 } from "../../utils/common";
-
-const major = [
-  {
-    item: "비서학과",
-    value: "1",
-  },
-  {
-    item: "관광서비스경영학과",
-    value: "2",
-  },
-  {
-    item: "휴먼사회복지학과",
-    value: "3",
-  },
-  {
-    item: "비지니스영어과",
-    value: "4",
-  },
-  {
-    item: "비즈니스중국어과",
-    value: "5",
-  },
-  {
-    item: "비지니스일본어과",
-    value: "6",
-  },
-  {
-    item: "세무회계학과",
-    value: "7",
-  },
-  {
-    item: "글로벌항공서비스학과",
-    value: "8",
-  },
-  {
-    item: "건축학과",
-    value: "9",
-  },
-  {
-    item: "토목공학과",
-    value: "10",
-  },
-  {
-    item: "실내건축과",
-    value: "11",
-  },
-  {
-    item: "디지털산업디자인학과",
-    value: "12",
-  },
-  {
-    item: "시각디자인과",
-    value: "13",
-  },
-  {
-    item: "주얼리디자인학과",
-    value: "14",
-  },
-  {
-    item: "멀티미디어디자인학과",
-    value: "15",
-  },
-  {
-    item: "정보통신공학과",
-    value: "16",
-  },
-  {
-    item: "리빙세라믹디자인학과",
-    value: "17",
-  },
-  {
-    item: "게임/vr디자인학과",
-    value: "18",
-  },
-  {
-    item: "방송영상미디어학과",
-    value: "19",
-  },
-  {
-    item: "방송뷰티학과",
-    value: "20",
-  },
-  {
-    item: "기계자동화학과",
-    value: "21",
-  },
-  {
-    item: "컴퓨터전자공학과",
-    value: "22",
-  },
-  {
-    item: "산업경영공학과",
-    value: "23",
-  },
-  {
-    item: "컴퓨터소프트웨어학과",
-    value: "24",
-  },
-  {
-    item: "메카트로닉스공학과",
-    value: "25",
-  },
-  {
-    item: "융합기계공학과",
-    value: "26",
-  },
-];
+import major from "../../utils/majors";
 
 const Container = styled.SafeAreaView`
   flex: 1;
@@ -246,72 +140,72 @@ function Signup() {
       <KeyboardAwareScrollView extraScrollHeight={30}>
         <Container>
           <Input
-            label="학번"
+            label='학번'
             value={student}
             onChangeText={(text) => setStudent(removeWhitespace(text))}
             onSubmitEditing={() => nameRef.current.focus()}
-            placeholder="학번"
-            returnKeyType="next"
+            placeholder='학번'
+            returnKeyType='next'
           />
           <SelectBox
-            label="학과"
+            label='학과'
             options={major}
-            inputPlaceholder="학과"
+            inputPlaceholder='학과'
             value={selectedMajor}
             onChange={onChange()}
             hideInputFilter={false}
           />
           <Input
-            label="이름"
+            label='이름'
             ref={nameRef}
             value={name}
             onChangeText={(text) => setName(removeWhitespace(text))}
             onSubmitEditing={() => nicknameRef.current.focus()}
-            placeholder="이름"
-            returnKeyType="next"
+            placeholder='이름'
+            returnKeyType='next'
           />
           <Input
-            label="별명"
+            label='별명'
             ref={nicknameRef}
             value={nickname}
             onChangeText={(text) => setNickname(removeWhitespace(text))}
             onSubmitEditing={() => emailRef.current.focus()}
-            placeholder="별명"
-            returnKeyType="next"
+            placeholder='별명'
+            returnKeyType='next'
           />
           <Input
-            label="이메일"
+            label='이메일'
             ref={emailRef}
             value={email}
             onChangeText={(text) => setEmail(removeWhitespace(text))}
             onSubmitEditing={() => passwordRef.current.focus()}
-            placeholder="이메일"
-            returnKeyType="next"
+            placeholder='이메일'
+            returnKeyType='next'
           />
           <Input
             ref={passwordRef}
-            label="비밀번호"
+            label='비밀번호'
             value={password}
             onChangeText={(text) => setPassword(removeWhitespace(text))}
             onSubmitEditing={() => passwordConfirmRef.current.focus()}
-            placeholder="비밀번호"
-            returnKeyType="next"
+            placeholder='비밀번호'
+            returnKeyType='next'
             isPassword
           />
           <Input
             ref={passwordConfirmRef}
-            label="비밀번호확인"
+            label='비밀번호확인'
             value={passwordConfirm}
             onChangeText={(text) => setPasswordConfirm(removeWhitespace(text))}
             onSubmitEditing={() => {}}
-            placeholder="비밀번호확인"
-            returnKeyType="done"
+            placeholder='비밀번호확인'
+            returnKeyType='done'
             isPassword
           />
           <ErrorText>{errorMessage}</ErrorText>
 
           <Button
-            title="회원가입"
+            title='회원가입'
             onPress={_handleSignupButtonPress}
             disabled={disabled}
           />
