+++ app/src/utils/majors.js
@@ -1,0 +1,108 @@
+const major = [
+  {
+    item: "비서학과",
+    value: "1",
+  },
+  {
+    item: "관광서비스경영학과",
+    value: "2",
+  },
+  {
+    item: "휴먼사회복지학과",
+    value: "3",
+  },
+  {
+    item: "비지니스영어과",
+    value: "4",
+  },
+  {
+    item: "비즈니스중국어과",
+    value: "5",
+  },
+  {
+    item: "비지니스일본어과",
+    value: "6",
+  },
+  {
+    item: "세무회계학과",
+    value: "7",
+  },
+  {
+    item: "글로벌항공서비스학과",
+    value: "8",
+  },
+  {
+    item: "건축학과",
+    value: "9",
+  },
+  {
+    item: "토목공학과",
+    value: "10",
+  },
+  {
+    item: "실내건축과",
+    value: "11",
+  },
+  {
+    item: "디지털산업디자인학과",
+    value: "12",
+  },
+  {
+    item: "시각디자인과",
+    value: "13",
+  },
+  {
+    item: "주얼리디자인학과",
+    value: "14",
+  },
+  {
+    item: "멀티미디어디자인학과",
+    value: "15",
+  },
+  {
+    item: "정보통신공학과",
+    value: "16",
+  },
+  {
+    item: "리빙세라믹디자인학과",
+    value: "17",
+  },
+  {
+    item: "게임/vr디자인학과",
+    value: "18",
+  },
+  {
+    item: "방송영상미디어학과",
+    value: "19",
+  },
+  {
+    item: "방송뷰티학과",
+    value: "20",
+  },
+  {
+    item: "기계자동화학과",
+    value: "21",
+  },
+  {
+    item: "컴퓨터전자공학과",
+    value: "22",
+  },
+  {
+    item: "산업경영공학과",
+    value: "23",
+  },
+  {
+    item: "컴퓨터소프트웨어학과",
+    value: "24",
+  },
+  {
+    item: "메카트로닉스공학과",
+    value: "25",
+  },
+  {
+    item: "융합기계공학과",
+    value: "26",
+  },
+];
+
+export default major;
